refactor: ChessWebController 추가
- WebUIChessApplication의 로직을 ChessWebController로  이동
- chess.html 추가
