+++ src/main/java/chess/WebUIChessApplication.java
@@ -1,22 +1,11 @@
 package chess;
 
-import static spark.Spark.*;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import spark.ModelAndView;
-import spark.template.handlebars.HandlebarsTemplateEngine;
+import chess.contoller.ChessWebController;
 
 public class WebUIChessApplication {
-    public static void main(String[] args) {
-        get("/", (req, res) -> {
-            Map<String, Object> model = new HashMap<>();
-            return render(model, "index.html");
-        });
-    }
+	public static void main(String[] args) {
+		ChessWebController chessWebController = new ChessWebController();
+		chessWebController.run();
+	}
 
-    private static String render(Map<String, Object> model, String templatePath) {
-        return new HandlebarsTemplateEngine().render(new ModelAndView(model, templatePath));
-    }
 }
