+++ src/main/resources/templates/chess.html
@@ -2,9 +2,13 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
-    <title>로또</title>
+    <title>Chess</title>
 </head>
 <body>
-<a href="/chess">게임 시작</a>
+<div>
+    <div>
+        Hi
+    </div>
+</div>
 </body>
 </html>
